requirements.txt
pyjwt==2.1.0
httpx==0.19.0
yarl==1.6.3